+++ R/PoolPrev.R
@@ -27,6 +27,11 @@
 #'   prior.
 #' @param level Defines the confidence level to be used for the confidence and
 #'   credible intervals. Defaults to 0.95 (i.e. 95\% intervals)
+#' @param reproduce.poolscreen (defaults to FALSE). If TRUE this changes the
+#'   way that likelihood ratio confidence intervals are computed to be somewhat
+#'   wider and more closely match those returned by Poolscreen. We recommend
+#'   using the default (FALSE). However setting to TRUE can help to make
+#'   comparisons between PoolPrev and Poolscreen.
 #' @param verbose Logical indicating whether to print progress to screen.
 #'   Defaults to false (no printing to screen).
 #' @param iter,warmup,chains MCMC options for passing onto the sampling
@@ -36,17 +41,23 @@
 #'   Defaults to default values as defined in \link[rstan]{stan}, except for
 #'   \code{adapt_delta} which is set to the more conservative value of 0.9. See
 #'   \link[rstan]{stan} for details.
-#' @return A \code{data.frame} with columns: \itemize{ \item{\code{PrevMLE} (the
-#'   Maximum Likelihood Estimate of prevalence)} \item{\code{CILow} and
-#'   \code{CIHigh} (Lower and Upper Confidence intervals using the Likelihood
-#'   Ratio method)} \item{\code{Bayesian Posterior Expectation}}
-#'   \item{\code{CrILow} and \code{CrIHigh}} \item{\code{Number of Pools}}
-#'   \item{\code{Number Positive}} } If grouping variables are provided in
-#'   \code{...} there will be an additional column for each grouping variable.
-#'   When there are no grouping variables (supplied in \code{...}) then the
-#'   dataframe has only one row with the prevalence estimates for the whole
-#'   dataset. When grouping variables are supplied, then there is a separate row
-#'   for each group.
+#' @return A \code{data.frame} with columns:
+#'   \itemize{\item{\code{PrevMLE} (the Maximum Likelihood Estimate of prevalence)}
+#'            \item{\code{CILow} and \code{CIHigh} - lower and upper confidence
+#'                  intervals using the likelihood ratio method}
+#'            \item{\code{PrevBayes} the (Bayesian) posterior expectation}
+#'            \item{\code{CrILow} and \code{CrIHigh} -- lower and upper bounds
+#'                  for credible intervals}
+#'            \item{\code{ProbAbsent} the posterior probability that prevalence
+#'                  is exactly 0 (i.e. disease marker is absent). NA if using
+#'                  default Jeffrey's prior or if prior.absent = 0.}
+#'            \item{\code{NumberOfPools} -- number of pools}
+#'            \item{\code{NumberPositive} -- the number of positive pools} }
+#'   If grouping variables are provided in \code{...} there will be an additional
+#'   column for each grouping variable. When there are no grouping variables
+#'   (supplied in \code{...}) then the output has only one row with the
+#'   prevalence estimates for the whole dataset. When grouping variables are
+#'   supplied, then there is a separate row for each group.
 #'
 #' @example examples/Prevalence.R
 #'      \code{\link{HierPoolPrev}},
@@ -55,7 +66,8 @@
 
 PoolPrev <- function(data,result,poolSize,...,
                      prior.alpha = NULL, prior.beta = NULL, prior.absent = 0,
-                     level = 0.95, verbose = FALSE,cores = NULL,
+                     level = 0.95, reproduce.poolscreen = FALSE,
+                     verbose = FALSE, cores = NULL,
                      iter = 2000, warmup = iter/2,
                      chains = 4, control = list(adapt_delta = 0.9)){
   result <- dplyr::enquo(result) #The name of column with the result of each test on each pooled sample
@@ -67,8 +79,6 @@ PoolPrev <- function(data,result,poolSize,...,
     stop("prior.alpha and prior.beta must either both be specified or both left blank. The latter uses the default Jeffrey's prior")
   }
 
-  # log-likelihood difference used to calculate Likelihood ratio confidence intervals
-  LogLikDiff <- stats::qchisq(level, df = 1)/2
   # log-likelihood function
   LogLikPrev = function(p,result,poolSize,goal=0){
     sum(log(result + (-1)^result * (1-p)^poolSize)) - goal
@@ -135,18 +145,17 @@ PoolPrev <- function(data,result,poolSize,...,
       MLEdata$JeffreysPrior  <- FALSE
       out$PrevMLE <- rstan::optimizing(stanmodels$BayesianPoolScreen,MLEdata)$par["p"]
 
-
+      # log-likelihood difference used to calculate Likelihood ratio confidence intervals
+      LogLikDiff <- stats::qchisq(if(reproduce.poolscreen){1 - (1 - level)/2}else{level}, df = 1)/2
       out[,'CILow'] <- stats::uniroot(LogLikPrev,
                                       c(0,out$PrevMLE),
-                                      goal = LogLikPrev(out$PrevMLE,sdata$Result,sdata$PoolSize) - LogLikDiff, # the version we would use if we let users supply confidence
-                                      #goal = LogLikPrev(out$PrevMLE,sdata$Result,sdata$PoolSize) - 2.51,
+                                      goal = LogLikPrev(out$PrevMLE,sdata$Result,sdata$PoolSize) - LogLikDiff,
                                       result= sdata$Result,
                                       poolSize= sdata$PoolSize,
                                       tol = 1e-10)$root
       out[,'CIHigh'] <- stats::uniroot(LogLikPrev,
                                        c(out$PrevMLE,1),
-                                       goal = LogLikPrev(out$PrevMLE,sdata$Result,sdata$PoolSize) - LogLikDiff, # the version we would use if we let users supply confidence
-                                       #goal = LogLikPrev(out$PrevMLE,sdata$Result,sdata$PoolSize) - 2.51, #Poolscreen uses the 2.51 value for a 95% confidence interval, which is the value one would use for 97.5% confidence interval (perhpas they thought they needed to make an adjustment for a 'two-sided' test?) For consistency I have reproduced it here
+                                       goal = LogLikPrev(out$PrevMLE,sdata$Result,sdata$PoolSize) - LogLikDiff,
                                        result= sdata$Result,
                                        poolSize= sdata$PoolSize,
                                        tol = 1e-10)$root
@@ -158,10 +167,10 @@ PoolPrev <- function(data,result,poolSize,...,
       out[,'CrIHigh'] <- 1
       out$ProbAbsent <- ifelse(prior.absent & !useJefferysPrior ,0,NA)
       out$PrevMLE <- 1
+      LogLikDiff <- stats::qchisq(level, df = 1)/2
       out[,'CILow'] <- stats::uniroot(LogLikPrev,
                                       c(0,1),
-                                      goal = -LogLikDiff, # the version we would use if we let users supply confidence
-                                      #goal = -1.92, # When all results are positive, the original Poolscreen uses this (more expected) value for the logliklihood difference (1.92) for a 95% confidence interval
+                                      goal = -LogLikDiff,
                                       result= sdata$Result,
                                       poolSize= sdata$PoolSize,
                                       tol = 1e-10)$root
@@ -191,10 +200,10 @@ PoolPrev <- function(data,result,poolSize,...,
       }
       out$PrevMLE <- 0
       out[,'CILow'] <- 0
+      LogLikDiff <- stats::qchisq(level, df = 1)/2
       out[,'CIHigh'] <- stats::uniroot(LogLikPrev,
                                        c(0,1),
-                                       goal = -LogLikDiff, # the version we would use if we let users supply confidence
-                                       #goal = -1.92, # When all results are negative, the original Poolscreen uses this (more expected) value for the logliklihood difference (1.92) for a 95% confidence interval
+                                       goal = -LogLikDiff,
                                        result= sdata$Result,
                                        poolSize= sdata$PoolSize,
                                        tol = 1e-10)$root
@@ -226,6 +235,7 @@ PoolPrev <- function(data,result,poolSize,...,
                prior.alpha = prior.alpha,
                prior.beta = prior.beta,
                prior.absent = prior.absent,
+               reproduce.poolscreen = reproduce.poolscreen,
                cores = cores,
                iter = iter,
                warmup = warmup,